Converted shortcode to a class
